+++ src/parsers/contest/UOJContestParser.ts
@@ -1,0 +1,12 @@
+import { ContestParser } from '../ContestParser';
+import { Parser } from '../Parser';
+import { UOJProblemParser } from '../problem/UOJProblemParser';
+
+export class UOJContestParser extends ContestParser {
+  public problemParser: Parser = new UOJProblemParser();
+  public linkSelector: string = '.top-buffer-md > .table-responsive > .table a';
+
+  public getMatchPatterns(): string[] {
+    return ['https://uoj.ac/contest/*', 'http://uoj.ac/contest/*'];
+  }
+}